refactor(helpers): consolidate template loading into loadTemplateByName (#370)
Follow-up to #369. SimpleAgentTemplatePage.load() and loadMemoryChatbot()
duplicated the same template-load sequence (delete all flows -> open templates
modal -> All Templates tab -> pick heading), so the #369 overlay fix had to be
applied in two bespoke paths.

- New shared helper `loadTemplateByName(page, name)` (tests/helpers/flows/).
  Deletes flows via the API (cleanAllFlows) instead of the home dropdown loop,
  which detached mid-click under re-renders and was a recurring flake source.
- Extended `openNewFlowTemplatesModal` to click the header button OR the
  empty-page CTA via `.or().first()`, so it covers both entry points and the
  inline `.or()` click no longer has to be duplicated per caller. Backward
  compatible with awaitBootstrapTest and bulk-actions (header is DOM-first when
  both render).
- Migrated SimpleAgentTemplatePage.load() and loadMemoryChatbot() onto the
  shared helper; they keep only their post-load steps (provider setup /
  updateOldComponents).

Net -22 lines. The bespoke UI deletion loop (btn_delete_dropdown_menu) now
exists in zero specs.

Validated green with --retries=0: memory x3, agent interaction suite, mcp-client
(migrated paths); api-request, bulk-actions, global-variables-crud (broad
openNewFlowTemplatesModal / awaitBootstrapTest paths, both empty and populated
home states).
